Split CarePAckage controller Organise use cases
