+++ BrokerageApi.Tests/V1/UseCase/CreateElementUseCaseTests.cs
@@ -9,6 +9,7 @@ using BrokerageApi.V1.Gateways.Interfaces;
 using BrokerageApi.V1.Infrastructure;
 using BrokerageApi.V1.Services.Interfaces;
 using BrokerageApi.V1.UseCase;
+using BrokerageApi.V1.UseCase.CarePackageElements;
 using FluentAssertions;
 using Moq;
 using NUnit.Framework;
