+++ BrokerageApi.Tests/V1/UseCase/DeleteElementUseCaseTests.cs
@@ -7,6 +7,7 @@ using BrokerageApi.Tests.V1.Helpers;
 using BrokerageApi.V1.Gateways.Interfaces;
 using BrokerageApi.V1.Infrastructure;
 using BrokerageApi.V1.Services.Interfaces;
+using BrokerageApi.V1.UseCase.CarePackageElements;
 using BrokerageApi.V1.UseCase.Interfaces;
 using FluentAssertions;
 using Moq;
