+++ BrokerageApi.Tests/V1/UseCase/GetCarePackageByIdUseCaseTests.cs
@@ -5,6 +5,7 @@ using BrokerageApi.Tests.V1.Helpers;
 using BrokerageApi.V1.Gateways.Interfaces;
 using BrokerageApi.V1.Infrastructure;
 using BrokerageApi.V1.UseCase;
+using BrokerageApi.V1.UseCase.CarePackages;
 using FluentAssertions;
 using Moq;
 using NUnit.Framework;
