+++ BrokerageApi.Tests/V1/UseCase/StartCarePackageUseCaseTests.cs
@@ -6,6 +6,7 @@ using BrokerageApi.V1.Gateways.Interfaces;
 using BrokerageApi.V1.Infrastructure;
 using BrokerageApi.V1.Services.Interfaces;
 using BrokerageApi.V1.UseCase;
+using BrokerageApi.V1.UseCase.CarePackages;
 using Moq;
 using NUnit.Framework;
 using NodaTime;
