+++ BrokerageApi.Tests/V1/UseCase/SuspendElementUseCaseTests.cs
@@ -11,6 +11,7 @@ using BrokerageApi.V1.Infrastructure.AuditEvents;
 using BrokerageApi.V1.Services;
 using BrokerageApi.V1.Services.Interfaces;
 using BrokerageApi.V1.UseCase;
+using BrokerageApi.V1.UseCase.CarePackageElements;
 using Castle.Core.Internal;
 using FluentAssertions;
 using Moq;
