+++ BrokerageApi/V1/UseCase/CarePackageElements/CreateElementUseCase.cs
@@ -7,9 +7,9 @@ using BrokerageApi.V1.Factories;
 using BrokerageApi.V1.Gateways.Interfaces;
 using BrokerageApi.V1.Infrastructure;
 using BrokerageApi.V1.Services.Interfaces;
-using BrokerageApi.V1.UseCase.Interfaces;
+using BrokerageApi.V1.UseCase.Interfaces.CarePackageElements;
 
-namespace BrokerageApi.V1.UseCase
+namespace BrokerageApi.V1.UseCase.CarePackageElements
 {
     public class CreateElementUseCase : ICreateElementUseCase
     {
