+++ BrokerageApi/V1/UseCase/CarePackageElements/DeleteElementUseCase.cs
@@ -3,8 +3,9 @@ using System.Threading.Tasks;
 using BrokerageApi.V1.Gateways.Interfaces;
 using BrokerageApi.V1.Infrastructure;
 using BrokerageApi.V1.Services.Interfaces;
+using BrokerageApi.V1.UseCase.Interfaces.CarePackageElements;
 
-namespace BrokerageApi.V1.UseCase.Interfaces
+namespace BrokerageApi.V1.UseCase.CarePackageElements
 {
     public class DeleteElementUseCase : IDeleteElementUseCase
     {
