+++ BrokerageApi/V1/UseCase/CarePackageElements/EndElementUseCase.cs
@@ -4,10 +4,10 @@ using BrokerageApi.V1.Gateways.Interfaces;
 using BrokerageApi.V1.Infrastructure;
 using BrokerageApi.V1.Infrastructure.AuditEvents;
 using BrokerageApi.V1.Services.Interfaces;
-using BrokerageApi.V1.UseCase.Interfaces;
+using BrokerageApi.V1.UseCase.Interfaces.CarePackageElements;
 using NodaTime;
 
-namespace BrokerageApi.V1.UseCase
+namespace BrokerageApi.V1.UseCase.CarePackageElements
 {
     public class EndElementUseCase : IEndElementUseCase
     {
