+++ BrokerageApi/V1/UseCase/CarePackages/StartCarePackageUseCase.cs
@@ -3,9 +3,9 @@ using System.Threading.Tasks;
 using BrokerageApi.V1.Gateways.Interfaces;
 using BrokerageApi.V1.Infrastructure;
 using BrokerageApi.V1.Services.Interfaces;
-using BrokerageApi.V1.UseCase.Interfaces;
+using BrokerageApi.V1.UseCase.Interfaces.CarePackages;
 
-namespace BrokerageApi.V1.UseCase
+namespace BrokerageApi.V1.UseCase.CarePackages
 {
     public class StartCarePackageUseCase : IStartCarePackageUseCase
     {
