+++ BrokerageApi/V1/UseCase/Interfaces/CarePackageElements/IEditElementUseCase.cs
@@ -2,7 +2,7 @@ using System.Threading.Tasks;
 using BrokerageApi.V1.Boundary.Request;
 using BrokerageApi.V1.Infrastructure;
 
-namespace BrokerageApi.V1.UseCase.Interfaces
+namespace BrokerageApi.V1.UseCase.Interfaces.CarePackageElements
 {
     public interface IEditElementUseCase
     {
