+++ BrokerageApi/V1/UseCase/Interfaces/CarePackageElements/IEndElementUseCase.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using NodaTime;
 
-namespace BrokerageApi.V1.UseCase.Interfaces
+namespace BrokerageApi.V1.UseCase.Interfaces.CarePackageElements
 {
     public interface IEndElementUseCase
     {
