+++ BrokerageApi/V1/UseCase/Interfaces/CarePackages/IStartCarePackageUseCase.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using BrokerageApi.V1.Infrastructure;
 
-namespace BrokerageApi.V1.UseCase.Interfaces
+namespace BrokerageApi.V1.UseCase.Interfaces.CarePackages
 {
     public interface IStartCarePackageUseCase
     {
